@@ -1,38 +1,74 @@
 {
-  "name": "body-parser",
-  "description": "Node.js body parsing middleware",
-  "version": "1.18.2",
-  "contributors": [
-    "Douglas Christopher Wilson <doug@somethingdoug.com>",
-    "Jonathan Ong <me@jongleberry.com> (http://jongleberry.com)"
+  "_from": "body-parser",
+  "_id": "body-parser@1.18.3",
+  "_inBundle": false,
+  "_integrity": "sha1-WykhmP/dVTs6DyDe0FkrlWlVyLQ=",
+  "_location": "/body-parser",
+  "_phantomChildren": {},
+  "_requested": {
+    "type": "tag",
+    "registry": true,
+    "raw": "body-parser",
+    "name": "body-parser",
+    "escapedName": "body-parser",
+    "rawSpec": "",
+    "saveSpec": null,
+    "fetchSpec": "latest"
+  },
+  "_requiredBy": [
+    "#USER",
+    "/",
+    "/express"
   ],
-  "license": "MIT",
-  "repository": "expressjs/body-parser",
+  "_resolved": "https://registry.npmjs.org/body-parser/-/body-parser-1.18.3.tgz",
+  "_shasum": "5b292198ffdd553b3a0f20ded0592b956955c8b4",
+  "_spec": "body-parser",
+  "_where": "/home/eugeniy/www/cpaMagnet/server",
+  "bugs": {
+    "url": "https://github.com/expressjs/body-parser/issues"
+  },
+  "bundleDependencies": false,
+  "contributors": [
+    {
+      "name": "Douglas Christopher Wilson",
+      "email": "doug@somethingdoug.com"
+    },
+    {
+      "name": "Jonathan Ong",
+      "email": "me@jongleberry.com",
+      "url": "http://jongleberry.com"
+    }
+  ],
   "dependencies": {
     "bytes": "3.0.0",
     "content-type": "~1.0.4",
     "debug": "2.6.9",
-    "depd": "~1.1.1",
-    "http-errors": "~1.6.2",
-    "iconv-lite": "0.4.19",
+    "depd": "~1.1.2",
+    "http-errors": "~1.6.3",
+    "iconv-lite": "0.4.23",
     "on-finished": "~2.3.0",
-    "qs": "6.5.1",
-    "raw-body": "2.3.2",
-    "type-is": "~1.6.15"
+    "qs": "6.5.2",
+    "raw-body": "2.3.3",
+    "type-is": "~1.6.16"
   },
+  "deprecated": false,
+  "description": "Node.js body parsing middleware",
   "devDependencies": {
-    "eslint": "3.19.0",
-    "eslint-config-standard": "10.2.1",
-    "eslint-plugin-import": "2.7.0",
+    "eslint": "4.19.1",
+    "eslint-config-standard": "11.0.0",
+    "eslint-plugin-import": "2.11.0",
     "eslint-plugin-markdown": "1.0.0-beta.6",
-    "eslint-plugin-node": "5.1.1",
-    "eslint-plugin-promise": "3.5.0",
-    "eslint-plugin-standard": "3.0.1",
+    "eslint-plugin-node": "6.0.1",
+    "eslint-plugin-promise": "3.7.0",
+    "eslint-plugin-standard": "3.1.0",
     "istanbul": "0.4.5",
     "methods": "1.1.2",
     "mocha": "2.5.3",
-    "safe-buffer": "5.1.1",
+    "safe-buffer": "5.1.2",
     "supertest": "1.1.0"
+  },
+  "engines": {
+    "node": ">= 0.8"
   },
   "files": [
     "lib/",
@@ -40,13 +76,18 @@
     "HISTORY.md",
     "index.js"
   ],
-  "engines": {
-    "node": ">= 0.8"
+  "homepage": "https://github.com/expressjs/body-parser#readme",
+  "license": "MIT",
+  "name": "body-parser",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/expressjs/body-parser.git"
   },
   "scripts": {
     "lint": "eslint --plugin markdown --ext js,md .",
     "test": "mocha --require test/support/env --reporter spec --check-leaks --bail test/",
     "test-cov": "istanbul cover node_modules/mocha/bin/_mocha -- --require test/support/env --reporter dot --check-leaks test/",
     "test-travis": "istanbul cover node_modules/mocha/bin/_mocha --report lcovonly -- --require test/support/env --reporter spec --check-leaks test/"
-  }
+  },
+  "version": "1.18.3"
 }
