@@ -1,23 +1,51 @@
 {
-  "name": "camelcase",
-  "version": "4.1.0",
-  "description": "Convert a dash/dot/underscore/space separated string to camelCase: foo-bar → fooBar",
-  "license": "MIT",
-  "repository": "sindresorhus/camelcase",
+  "_from": "camelcase@^5.0.0",
+  "_id": "camelcase@5.3.1",
+  "_inBundle": false,
+  "_integrity": "sha512-L28STB170nwWS63UjtlEOE3dldQApaJXZkOI1uMFfzf3rRuPegHaHesyee+YxQ+W6SvRDQV6UrdOdRiR153wJg==",
+  "_location": "/camelcase",
+  "_phantomChildren": {},
+  "_requested": {
+    "type": "range",
+    "registry": true,
+    "raw": "camelcase@^5.0.0",
+    "name": "camelcase",
+    "escapedName": "camelcase",
+    "rawSpec": "^5.0.0",
+    "saveSpec": null,
+    "fetchSpec": "^5.0.0"
+  },
+  "_requiredBy": [
+    "/yargs-parser"
+  ],
+  "_resolved": "https://registry.npmjs.org/camelcase/-/camelcase-5.3.1.tgz",
+  "_shasum": "e3c9b31569e106811df242f715725a1f4c494320",
+  "_spec": "camelcase@^5.0.0",
+  "_where": "/home/eugeniy/www/cpaMagnet/server/node_modules/yargs-parser",
   "author": {
     "name": "Sindre Sorhus",
     "email": "sindresorhus@gmail.com",
     "url": "sindresorhus.com"
   },
-  "engines": {
-    "node": ">=4"
+  "bugs": {
+    "url": "https://github.com/sindresorhus/camelcase/issues"
   },
-  "scripts": {
-    "test": "xo && ava"
+  "bundleDependencies": false,
+  "deprecated": false,
+  "description": "Convert a dash/dot/underscore/space separated string to camelCase or PascalCase: `foo-bar` → `fooBar`",
+  "devDependencies": {
+    "ava": "^1.4.1",
+    "tsd": "^0.7.1",
+    "xo": "^0.24.0"
+  },
+  "engines": {
+    "node": ">=6"
   },
   "files": [
-    "index.js"
+    "index.js",
+    "index.d.ts"
   ],
+  "homepage": "https://github.com/sindresorhus/camelcase#readme",
   "keywords": [
     "camelcase",
     "camel-case",
@@ -30,13 +58,18 @@
     "separator",
     "string",
     "text",
-    "convert"
+    "convert",
+    "pascalcase",
+    "pascal-case"
   ],
-  "devDependencies": {
-    "ava": "*",
-    "xo": "*"
+  "license": "MIT",
+  "name": "camelcase",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/sindresorhus/camelcase.git"
   },
-  "xo": {
-    "esnext": true
-  }
+  "scripts": {
+    "test": "xo && ava && tsd"
+  },
+  "version": "5.3.1"
 }
