@@ -1,31 +1,78 @@
 {
-  "name": "cors",
-  "description": "Node.js CORS middleware",
-  "version": "2.8.4",
-  "author": "Troy Goode <troygoode@gmail.com> (https://github.com/troygoode/)",
-  "keywords": ["cors", "express", "connect", "middleware"],
-  "license": "MIT",
-  "repository": "expressjs/cors",
-  "main": "./lib/index.js",
-  "engines": {
-    "node": ">=0.10.0"
+  "_from": "cors",
+  "_id": "cors@2.8.5",
+  "_inBundle": false,
+  "_integrity": "sha512-KIHbLJqu73RGr/hnbrO9uBeixNGuvSQjul/jdFvS/KFSIH1hWVd1ng7zOHx+YrEfInLG7q4n6GHQ9cDtxv/P6g==",
+  "_location": "/cors",
+  "_phantomChildren": {},
+  "_requested": {
+    "type": "tag",
+    "registry": true,
+    "raw": "cors",
+    "name": "cors",
+    "escapedName": "cors",
+    "rawSpec": "",
+    "saveSpec": null,
+    "fetchSpec": "latest"
   },
+  "_requiredBy": [
+    "#USER",
+    "/"
+  ],
+  "_resolved": "https://registry.npmjs.org/cors/-/cors-2.8.5.tgz",
+  "_shasum": "eac11da51592dd86b9f06f6e7ac293b3df875d29",
+  "_spec": "cors",
+  "_where": "/home/eugeniy/www/cpaMagnet/server",
+  "author": {
+    "name": "Troy Goode",
+    "email": "troygoode@gmail.com",
+    "url": "https://github.com/troygoode/"
+  },
+  "bugs": {
+    "url": "https://github.com/expressjs/cors/issues"
+  },
+  "bundleDependencies": false,
   "dependencies": {
     "object-assign": "^4",
     "vary": "^1"
   },
+  "deprecated": false,
+  "description": "Node.js CORS middleware",
   "devDependencies": {
-    "basic-auth-connect": "^1.0.0",
-    "body-parser": "^1.12.4",
-    "eslint": "^0.21.2",
-    "express": "^4.12.4",
-    "istanbul": "^0.4.5",
-    "mocha": "3.4.2",
-    "should": "11.2.1",
-    "supertest": "3.0.0"
+    "after": "0.8.2",
+    "eslint": "2.13.1",
+    "express": "4.16.3",
+    "mocha": "5.2.0",
+    "nyc": "13.1.0",
+    "supertest": "3.3.0"
+  },
+  "engines": {
+    "node": ">= 0.10"
+  },
+  "files": [
+    "lib/index.js",
+    "CONTRIBUTING.md",
+    "HISTORY.md",
+    "LICENSE",
+    "README.md"
+  ],
+  "homepage": "https://github.com/expressjs/cors#readme",
+  "keywords": [
+    "cors",
+    "express",
+    "connect",
+    "middleware"
+  ],
+  "license": "MIT",
+  "main": "./lib/index.js",
+  "name": "cors",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/expressjs/cors.git"
   },
   "scripts": {
-    "test": "npm run lint && istanbul cover node_modules/mocha/bin/_mocha",
-    "lint": "eslint lib test"
-  }
+    "lint": "eslint lib test",
+    "test": "npm run lint && nyc --reporter=html --reporter=text mocha --require test/support/env"
+  },
+  "version": "2.8.5"
 }
