@@ -1,31 +1,65 @@
 {
-  "name": "execa",
-  "version": "0.7.0",
-  "description": "A better `child_process`",
-  "license": "MIT",
-  "repository": "sindresorhus/execa",
+  "_from": "execa@^1.0.0",
+  "_id": "execa@1.0.0",
+  "_inBundle": false,
+  "_integrity": "sha512-adbxcyWV46qiHyvSp50TKt05tB4tK3HcmF7/nxfAdhnox83seTDbwnaqKO4sXRy7roHAIFqJP/Rw/AuEbX61LA==",
+  "_location": "/execa",
+  "_phantomChildren": {},
+  "_requested": {
+    "type": "range",
+    "registry": true,
+    "raw": "execa@^1.0.0",
+    "name": "execa",
+    "escapedName": "execa",
+    "rawSpec": "^1.0.0",
+    "saveSpec": null,
+    "fetchSpec": "^1.0.0"
+  },
+  "_requiredBy": [
+    "/os-locale"
+  ],
+  "_resolved": "https://registry.npmjs.org/execa/-/execa-1.0.0.tgz",
+  "_shasum": "c6236a5bb4df6d6f15e88e7f017798216749ddd8",
+  "_spec": "execa@^1.0.0",
+  "_where": "/home/eugeniy/www/cpaMagnet/server/node_modules/os-locale",
   "author": {
     "name": "Sindre Sorhus",
     "email": "sindresorhus@gmail.com",
     "url": "sindresorhus.com"
   },
-  "maintainers": [
-    {
-      "name": "James Talmage",
-      "email": "james@talmage.io",
-      "url": "github.com/jamestalmage"
-    }
-  ],
-  "engines": {
-    "node": ">=4"
+  "bugs": {
+    "url": "https://github.com/sindresorhus/execa/issues"
   },
-  "scripts": {
-    "test": "xo && nyc ava"
+  "bundleDependencies": false,
+  "dependencies": {
+    "cross-spawn": "^6.0.0",
+    "get-stream": "^4.0.0",
+    "is-stream": "^1.1.0",
+    "npm-run-path": "^2.0.0",
+    "p-finally": "^1.0.0",
+    "signal-exit": "^3.0.0",
+    "strip-eof": "^1.0.0"
+  },
+  "deprecated": false,
+  "description": "A better `child_process`",
+  "devDependencies": {
+    "ava": "*",
+    "cat-names": "^1.0.2",
+    "coveralls": "^3.0.1",
+    "delay": "^3.0.0",
+    "is-running": "^2.0.0",
+    "nyc": "^13.0.1",
+    "tempfile": "^2.0.0",
+    "xo": "*"
+  },
+  "engines": {
+    "node": ">=6"
   },
   "files": [
     "index.js",
     "lib"
   ],
+  "homepage": "https://github.com/sindresorhus/execa#readme",
   "keywords": [
     "exec",
     "child",
@@ -43,25 +77,8 @@
     "path",
     "local"
   ],
-  "dependencies": {
-    "cross-spawn": "^5.0.1",
-    "get-stream": "^3.0.0",
-    "is-stream": "^1.1.0",
-    "npm-run-path": "^2.0.0",
-    "p-finally": "^1.0.0",
-    "signal-exit": "^3.0.0",
-    "strip-eof": "^1.0.0"
-  },
-  "devDependencies": {
-    "ava": "*",
-    "cat-names": "^1.0.2",
-    "coveralls": "^2.11.9",
-    "delay": "^2.0.0",
-    "is-running": "^2.0.0",
-    "nyc": "^11.0.2",
-    "tempfile": "^2.0.0",
-    "xo": "*"
-  },
+  "license": "MIT",
+  "name": "execa",
   "nyc": {
     "reporter": [
       "text",
@@ -72,5 +89,13 @@
       "**/test.js",
       "**/test/**"
     ]
-  }
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/sindresorhus/execa.git"
+  },
+  "scripts": {
+    "test": "xo && nyc ava"
+  },
+  "version": "1.0.0"
 }
