@@ -1,31 +1,60 @@
 {
-  "name": "get-stream",
-  "version": "3.0.0",
-  "description": "Get a stream as a string, buffer, or array",
-  "license": "MIT",
-  "repository": "sindresorhus/get-stream",
+  "_from": "get-stream@^4.0.0",
+  "_id": "get-stream@4.1.0",
+  "_inBundle": false,
+  "_integrity": "sha512-GMat4EJ5161kIy2HevLlr4luNjBgvmj413KaQA7jt4V8B4RDsfpHk7WQ9GVqfYyyx8OS/L66Kox+rJRNklLK7w==",
+  "_location": "/get-stream",
+  "_phantomChildren": {},
+  "_requested": {
+    "type": "range",
+    "registry": true,
+    "raw": "get-stream@^4.0.0",
+    "name": "get-stream",
+    "escapedName": "get-stream",
+    "rawSpec": "^4.0.0",
+    "saveSpec": null,
+    "fetchSpec": "^4.0.0"
+  },
+  "_requiredBy": [
+    "/execa"
+  ],
+  "_resolved": "https://registry.npmjs.org/get-stream/-/get-stream-4.1.0.tgz",
+  "_shasum": "c1b255575f3dc21d59bfc79cd3d2b46b1c3a54b5",
+  "_spec": "get-stream@^4.0.0",
+  "_where": "/home/eugeniy/www/cpaMagnet/server/node_modules/execa",
   "author": {
     "name": "Sindre Sorhus",
     "email": "sindresorhus@gmail.com",
     "url": "sindresorhus.com"
   },
-  "engines": {
-    "node": ">=4"
+  "bugs": {
+    "url": "https://github.com/sindresorhus/get-stream/issues"
   },
-  "scripts": {
-    "test": "xo && ava"
+  "bundleDependencies": false,
+  "dependencies": {
+    "pump": "^3.0.0"
+  },
+  "deprecated": false,
+  "description": "Get a stream as a string, buffer, or array",
+  "devDependencies": {
+    "ava": "*",
+    "into-stream": "^3.0.0",
+    "xo": "*"
+  },
+  "engines": {
+    "node": ">=6"
   },
   "files": [
     "index.js",
     "buffer-stream.js"
   ],
+  "homepage": "https://github.com/sindresorhus/get-stream#readme",
   "keywords": [
     "get",
     "stream",
     "promise",
     "concat",
     "string",
-    "str",
     "text",
     "buffer",
     "read",
@@ -34,15 +63,16 @@
     "readable",
     "readablestream",
     "array",
-    "object",
-    "obj"
+    "object"
   ],
-  "devDependencies": {
-    "ava": "*",
-    "into-stream": "^3.0.0",
-    "xo": "*"
+  "license": "MIT",
+  "name": "get-stream",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/sindresorhus/get-stream.git"
   },
-  "xo": {
-    "esnext": true
-  }
+  "scripts": {
+    "test": "xo && ava"
+  },
+  "version": "4.1.0"
 }
