@@ -1,32 +1,52 @@
 {
-  "name": "statuses",
-  "description": "HTTP status utility",
-  "version": "1.4.0",
+  "_from": "statuses@>= 1.4.0 < 2",
+  "_id": "statuses@1.5.0",
+  "_inBundle": false,
+  "_integrity": "sha1-Fhx9rBd2Wf2YEfQ3cfqZOBR4Yow=",
+  "_location": "/statuses",
+  "_phantomChildren": {},
+  "_requested": {
+    "type": "range",
+    "registry": true,
+    "raw": "statuses@>= 1.4.0 < 2",
+    "name": "statuses",
+    "escapedName": "statuses",
+    "rawSpec": ">= 1.4.0 < 2",
+    "saveSpec": null,
+    "fetchSpec": ">= 1.4.0 < 2"
+  },
+  "_requiredBy": [
+    "/http-errors"
+  ],
+  "_resolved": "https://registry.npmjs.org/statuses/-/statuses-1.5.0.tgz",
+  "_shasum": "161c7dac177659fd9811f43771fa99381478628c",
+  "_spec": "statuses@>= 1.4.0 < 2",
+  "_where": "/home/eugeniy/www/cpaMagnet/server/node_modules/http-errors",
+  "bugs": {
+    "url": "https://github.com/jshttp/statuses/issues"
+  },
+  "bundleDependencies": false,
   "contributors": [
-    "Douglas Christopher Wilson <doug@somethingdoug.com>",
-    "Jonathan Ong <me@jongleberry.com> (http://jongleberry.com)"
+    {
+      "name": "Douglas Christopher Wilson",
+      "email": "doug@somethingdoug.com"
+    },
+    {
+      "name": "Jonathan Ong",
+      "email": "me@jongleberry.com",
+      "url": "http://jongleberry.com"
+    }
   ],
-  "repository": "jshttp/statuses",
-  "license": "MIT",
-  "keywords": [
-    "http",
-    "status",
-    "code"
-  ],
-  "files": [
-    "HISTORY.md",
-    "index.js",
-    "codes.json",
-    "LICENSE"
-  ],
+  "deprecated": false,
+  "description": "HTTP status utility",
   "devDependencies": {
     "csv-parse": "1.2.4",
-    "eslint": "3.19.0",
-    "eslint-config-standard": "10.2.1",
-    "eslint-plugin-import": "2.8.0",
+    "eslint": "4.19.1",
+    "eslint-config-standard": "11.0.0",
+    "eslint-plugin-import": "2.9.0",
     "eslint-plugin-markdown": "1.0.0-beta.6",
-    "eslint-plugin-node": "5.2.0",
-    "eslint-plugin-promise": "3.6.0",
+    "eslint-plugin-node": "6.0.1",
+    "eslint-plugin-promise": "3.7.0",
     "eslint-plugin-standard": "3.0.1",
     "istanbul": "0.4.5",
     "mocha": "1.21.5",
@@ -36,6 +56,24 @@
   "engines": {
     "node": ">= 0.6"
   },
+  "files": [
+    "HISTORY.md",
+    "index.js",
+    "codes.json",
+    "LICENSE"
+  ],
+  "homepage": "https://github.com/jshttp/statuses#readme",
+  "keywords": [
+    "http",
+    "status",
+    "code"
+  ],
+  "license": "MIT",
+  "name": "statuses",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/jshttp/statuses.git"
+  },
   "scripts": {
     "build": "node scripts/build.js",
     "fetch": "node scripts/fetch-apache.js && node scripts/fetch-iana.js && node scripts/fetch-nginx.js && node scripts/fetch-node.js",
@@ -44,5 +82,6 @@
     "test-ci": "istanbul cover node_modules/mocha/bin/_mocha --report lcovonly -- --reporter spec --check-leaks test/",
     "test-cov": "istanbul cover node_modules/mocha/bin/_mocha -- --reporter dot --check-leaks test/",
     "update": "npm run fetch && npm run build"
-  }
+  },
+  "version": "1.5.0"
 }
