@@ -1,55 +1,88 @@
 {
-	"name": "string-width",
-	"version": "2.1.1",
-	"description": "Get the visual width of a string - the number of columns required to display it",
-	"license": "MIT",
-	"repository": "sindresorhus/string-width",
-	"author": {
-		"name": "Sindre Sorhus",
-		"email": "sindresorhus@gmail.com",
-		"url": "sindresorhus.com"
-	},
-	"engines": {
-		"node": ">=4"
-	},
-	"scripts": {
-		"test": "xo && ava"
-	},
-	"files": [
-		"index.js"
-	],
-	"keywords": [
-		"string",
-		"str",
-		"character",
-		"char",
-		"unicode",
-		"width",
-		"visual",
-		"column",
-		"columns",
-		"fullwidth",
-		"full-width",
-		"full",
-		"ansi",
-		"escape",
-		"codes",
-		"cli",
-		"command-line",
-		"terminal",
-		"console",
-		"cjk",
-		"chinese",
-		"japanese",
-		"korean",
-		"fixed-width"
-	],
-	"dependencies": {
-		"is-fullwidth-code-point": "^2.0.0",
-		"strip-ansi": "^4.0.0"
-	},
-	"devDependencies": {
-		"ava": "*",
-		"xo": "*"
-	}
+  "_from": "string-width@^3.0.0",
+  "_id": "string-width@3.1.0",
+  "_inBundle": false,
+  "_integrity": "sha512-vafcv6KjVZKSgz06oM/H6GDBrAtz8vdhQakGjFIvNrHA6y3HCF1CInLy+QLq8dTJPQ1b+KDUqDFctkdRW44e1w==",
+  "_location": "/string-width",
+  "_phantomChildren": {},
+  "_requested": {
+    "type": "range",
+    "registry": true,
+    "raw": "string-width@^3.0.0",
+    "name": "string-width",
+    "escapedName": "string-width",
+    "rawSpec": "^3.0.0",
+    "saveSpec": null,
+    "fetchSpec": "^3.0.0"
+  },
+  "_requiredBy": [
+    "/yargs"
+  ],
+  "_resolved": "https://registry.npmjs.org/string-width/-/string-width-3.1.0.tgz",
+  "_shasum": "22767be21b62af1081574306f69ac51b62203961",
+  "_spec": "string-width@^3.0.0",
+  "_where": "/home/eugeniy/www/cpaMagnet/server/node_modules/yargs",
+  "author": {
+    "name": "Sindre Sorhus",
+    "email": "sindresorhus@gmail.com",
+    "url": "sindresorhus.com"
+  },
+  "bugs": {
+    "url": "https://github.com/sindresorhus/string-width/issues"
+  },
+  "bundleDependencies": false,
+  "dependencies": {
+    "emoji-regex": "^7.0.1",
+    "is-fullwidth-code-point": "^2.0.0",
+    "strip-ansi": "^5.1.0"
+  },
+  "deprecated": false,
+  "description": "Get the visual width of a string - the number of columns required to display it",
+  "devDependencies": {
+    "ava": "^1.0.1",
+    "xo": "^0.23.0"
+  },
+  "engines": {
+    "node": ">=6"
+  },
+  "files": [
+    "index.js"
+  ],
+  "homepage": "https://github.com/sindresorhus/string-width#readme",
+  "keywords": [
+    "string",
+    "str",
+    "character",
+    "char",
+    "unicode",
+    "width",
+    "visual",
+    "column",
+    "columns",
+    "fullwidth",
+    "full-width",
+    "full",
+    "ansi",
+    "escape",
+    "codes",
+    "cli",
+    "command-line",
+    "terminal",
+    "console",
+    "cjk",
+    "chinese",
+    "japanese",
+    "korean",
+    "fixed-width"
+  ],
+  "license": "MIT",
+  "name": "string-width",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/sindresorhus/string-width.git"
+  },
+  "scripts": {
+    "test": "xo && ava"
+  },
+  "version": "3.1.0"
 }
