@@ -1,69 +1,66 @@
-const express = require('express');
-const cors = require('cors');
-const graphqlHTTP = require('express-graphql');
-const gql = require('graphql-tag');
-const { buildASTSchema } = require('graphql');
+import fs from 'fs';
+import path from 'path';
+import cors from 'cors';
+import yargs from 'yargs';
+import express from 'express';
+import bodyParser from 'body-parser';
+import cookieParser from 'cookie-parser';
+import methodOverride from 'method-override';
+import promise from 'bluebird';
+import jwt from 'express-jwt';
+import config from 'config';
 
-const POSTS = [
-  { author: 'John Doe', body: 'Hello world' },
-  { author: 'Jane Doe', body: 'Hi, planet!' }
-];
+const argv = yargs
+  .default('port', config.get('server.port'))
+  .default('host', config.get('server.host'))
+  .alias('p', 'port')
+  .alias('h', 'host').argv;
 
-const schema = buildASTSchema(gql`
-  type Mutation {
-    submitPost(input: PostInput!): Post
-  }
-
-  input PostInput {
-    id: ID
-    author: String!
-    body: String!
-  }
-  type Query {
-    posts: [Post]
-    post(id: ID!): Post
-  }
-
-  type Post {
-    id: ID
-    author: String
-    body: String
-  }
-`);
-
-const mapPost = (post, id) => post && { id, ...post };
-
-const root = {
-  posts: () => POSTS.map(mapPost),
-  post: ({ id }) => mapPost(POSTS[id], id),
-  submitPost: ({ input: { id, author, body } }) => {
-    const post = { author, body };
-    let index = POSTS.length;
-
-    if (id != null && id >= 0 && id < POSTS.length) {
-      if (POSTS[id].authorId !== authorId) return null;
-
-      POSTS.splice(id, 1, post);
-      index = id;
-    } else {
-      POSTS.push(post);
-    }
-
-    return mapPost(post, index);
-  }
-};
+const db = require('./models')();
 
 const app = express();
-app.use(cors());
+
 app.use(
-  '/graphql',
-  graphqlHTTP({
-    schema,
-    rootValue: root,
-    graphiql: true
+  cors({
+    origin: '*',
+    methods: 'GET,HEAD,PUT,PATCH,POST,DELETE',
+    preflightContinue: false,
+    optionsSuccessStatus: 204,
+    credentials: true
   })
 );
+app.use(bodyParser.json({ limit: '100mb' }));
+app.use(bodyParser.urlencoded({ extended: false }));
+app.use(cookieParser());
+app.use(methodOverride('X-HTTP-Method-Override'));
 
-const port = process.env.PORT || 4000;
-app.listen(port);
-console.log(`Running a GraphQL API server at localhost:${port}/graphql`);
+app.use(express.static('public'));
+// app.use(require('./jwtMiddleware')());
+// app.use(
+//   jwt({
+//     secret: config.get('jwt.secretKey'),
+//     credentialsRequired: false,
+//     getToken: req => req.headers['x-access-token'] || null,
+//   }).unless({ path: ['/api/v1/login'] })
+// );
+
+require('./resources')(app, db);
+
+app.use(function errorHandler(err, req, res, next) {
+  if (res.headersSent) {
+    return next(err);
+  }
+
+  console.error(err);
+  res.status(500).json({ message: 'Unexpected error occurred' });
+});
+
+process.on('uncaughtException', err => {
+  console.error(err);
+});
+
+app.listen(argv.port, argv.host, () => {
+  console.log(
+    `${config.get('env')} server started at ${argv.host}: ${argv.port}`
+  );
+});
